feat(hardware): add sensor binding context manager
Taking actions while specific sensors are bound to an output is needed
for a couple different features of the OT3. Since only one sensor can be
bound to an output at a time, it's important for that binding to end
when it's relevant to do so. The best way to do that is a context
manager, which this adds: in
opentrons_hardware.sensors.scheduler.Scheduler, you can now do

async with scheduler.bind_sync(...):

to get an async context manager that binds a sensor and cleans up
afterwards.

This also required extending the rest of the scheduler functionality a
bit more - waiting for a new kind of response - which meant it made
sense to do a slight refactor of the response waiter to properly format
data. There were also some dataclass changes that required adapting
other classes.